refac: add package exception and middleware
