+++ internal/domain/permissions.go
@@ -1,5 +1,7 @@
 package domain
 
+import "context"
+
 type Permission string
 
 const (
