+++ pkg/server/routes.go
@@ -7,6 +7,8 @@ import (
 	"net/http"
 
 	"github.com/yawnak/foodadvisor/internal/domain"
+	"github.com/yawnak/foodadvisor/pkg/server/exception"
+	"github.com/yawnak/foodadvisor/pkg/server/middleware"
 )
 
 func (srv *Server) initAPIRoutes() {
@@ -14,8 +16,8 @@ func (srv *Server) initAPIRoutes() {
 	api := srv.router.PathPrefix("/api").Subrouter()
 
 	//sign up endpoints
-	api.HandleFunc("/signup", validateContentJSON(srv.signup)).Methods("POST")
-	api.HandleFunc("/login", validateContentJSON(srv.login)).Methods("POST")
+	api.HandleFunc("/signup", middleware.ValidateContentJSON(srv.signup)).Methods("POST")
+	api.HandleFunc("/login", middleware.ValidateContentJSON(srv.login)).Methods("POST")
 
 	// /api/users/ routes
 	//users := api.PathPrefix("/users").Subrouter()
@@ -24,20 +26,20 @@ func (srv *Server) initAPIRoutes() {
 	usersAuth.Use(srv.authenticate)
 
 	usersAuthAndPerms := usersAuth.PathPrefix("").Subrouter()
-	usersAuthAndPerms.Use(confirmPermissions(domain.PermEditUserRole))
+	usersAuthAndPerms.Use(middleware.ConfirmPermissions(domain.PermEditUserRole))
 	usersAuthAndPerms.HandleFunc("/{id:[0-9]+}/role", srv.setUserRole).Methods("POST")
 
 	roles := api.PathPrefix("/roles").Subrouter()
 	roles.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		role, ok := retrieveRole(r.Context())
+		role, ok := middleware.RetrieveRole(r.Context())
 		log.Println("ROLES")
 		if !ok {
-			writeErrorAsJSON(w, http.StatusInternalServerError, errors.New("not authorized"))
+			exception.WriteErrorAsJSON(w, http.StatusInternalServerError, errors.New("not authorized"))
 			log.Fatalln("didn't work. no role in context")
 			return
 		}
 		fmt.Fprintf(w, "hello user with role: %s", role.Name)
 	}).Methods("GET")
 	roles.Use(srv.authenticate)
-	roles.Use(confirmPermissions(domain.PermEditRoles))
+	roles.Use(middleware.ConfirmPermissions(domain.PermEditRoles))
 }
